refactor: 타입 정의들 각 entity 내의 model 디렉토리에 정의
