+++ src/entities/comment/api/deleteComment.ts
@@ -1,4 +1,4 @@
-import { Comment } from "../../../pages/PostsManagerPage"
+import { Comment } from "../model"
 
 interface DeleteCommentRequest {
   commentId: number
