+++ src/entities/post/api/deletePost.ts
@@ -1,4 +1,4 @@
-import { Post } from "../../../pages/PostsManagerPage"
+import { Post } from "../model"
 
 interface DeletePostRequest {
   postId: number
