+++ src/entities/post/api/fetchPostsBySearch.ts
@@ -1,4 +1,4 @@
-import { Post } from "../../../pages/PostsManagerPage"
+import { Post } from "../model"
 
 interface FetchPostsByQueryRequest {
   searchQuery: string
