+++ src/entities/post/api/updatePost.ts
@@ -1,4 +1,4 @@
-import { Post } from "../../../pages/PostsManagerPage"
+import { Post } from "../model"
 
 interface UpdatePostRequest {
   selectedPost: Post
